Replced logic for atomic orders from ui to ws
